:truck: Moved network folder into src
