+++ src/network/CMakeLists.txt
@@ -27,14 +27,14 @@ set(CMAKE_RUNTIME_OUTPUT_DIRECTORY_RELEASE ${CMAKE_SOURCE_DIR})
 
 
 # Add an executable for the server project
-add_executable(udp_server
-    udp_server.cpp
-    UDPServer.cpp
-)
+# add_executable(udp_server
+#     udp_server.cpp
+#     UDPServer.cpp
+# )
 
-target_link_libraries(udp_server
-    Boost::asio
-)
+# target_link_libraries(udp_server
+#     Boost::asio
+# )
 
 add_executable(udp_client
     udp_client.cpp
